Refactor PriceLevel and PhotoGallery to be components.
