+++ client/src/pages/SearchResult/SingleSearchResultStyles.tsx
@@ -16,7 +16,6 @@ export const Wrapper = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  background-color: black;
 `;
 
 export const HeartIcon = styled.img`
